Read TYPELAMBDAtype and PARAMtype, which refers to a parameter of a type lambda
The result of a type lambda refers to its parameters and the parameters of a type
lambda are defined by the type lambda and the number in the parameter list. To
break the cyclic reference, the result type of a lambda is read in the constructor,
when the type lambda object already exists and can be used to construct references
to the parameters. Therefore a type lambda is defined by the parameter list and
a function which reads the result type, given the type lambda object.