Generalize reading parameters of a defdef
The previous assumption of the unpickler was that the parameters of a method are
always a clause of type parameters, followed by 0, 1 or several clauses of
parameters. This is not true for generic extension methods, which have
param (extended class) - type params - other params. Therefore keep all parameters
of a DefDef in a single list of parameter clauses, as suggested by TASTy grammar.