Add a workaround for packages, nested in the empty package
By definition, empty package cannot contain any nested packages. However, for
the following code:

import something

package apackage {
    ...
}

the generated TASTy file represents apackage as being nested in empty package.
Therefore, the unpickler has to correct it and change the owner to root when a
package symbol is created in the empty package.