Give types to all the simple things.
This requires synthetic declarations for `scala.Nothing` and
`scala.Null`. They are added when we initialize the root context.

Applications do not receive types yet, because that will require
`calculateType` to actually resolve symbols to get to functions'
`MethodType`s.

In the process, we rename and fix SelectTypeFromTerm to
TermRefTypeTree, which selects a term ref from a prefix.