Refactor tests: make the intent clear when we want unforcing tests.
Most tests should check the surface API, and therefore the normal
ability of the API to lazy-load symbols on demand. Some rare tests
are designed to check that things are not loaded earlier than
needed.

We rename and give a better signature to the methods that are
meant for the latter kind of tests. In particular, we give them
long names that make it clear that they test the current internal
state.

For most tests, we use the `resolve` method. The intent of that
method is that it will load things as required. Currently, it does
not do that yet. This is left for a subsequent commit.