Make default symbol/decl lookup load things when necessary.
The previous behavior of finding decls without loading anything
was renamed as `getDeclInternal`. It is used in a couple of
internal locations where we really have no business loading stuff,
because we are already within code that loads.

Otherwise, methods that lookup symbols require an implicit
`BaseContext` so that they can load things when necessary.

This allows to remove the hacks in the TypeSuite to load symbols.
They get loaded automatically.