Simplify the testing infrastructure.
The overall test classpath now always contains everything.

We decompose BaseUnpicklingSuite into an unrestricted one and a
restricted one. `UnrestrictedUnpicklingSuite` gives access to the
entire classpath, and therefore can create contexts without having
to specify a top-level class. `RestrictedUnpicklingSuite` gives
tools to restrict the classpath to a single class, or a selected
subset of classes.