Remove ClassContext and {class,module,moduleClass}Root's.
We use `Context` everywhere instead of `ClassContext`. Where we
need an `owner`, we pass it down explicitly.

We get rid of the concept of class/module/moduleClass roots. In
Scala 2 pickles, roots are just normal classes and modules, which
receive no special treatment. In Java classfiles, the roots are
eagerly created in `loadJavaClass`.