Make FileContext private to TreeUnpickler.
Instead of being a `Context` itself, it is now accompanied by a
`Context`. Since there is now a unique `Context`, we can store it
in `TreeUnpickler` itself, rather than pass it to every method.

The resulting `FileContext` only contains *local* information, and
hence we rename it to `LocalContext`.

This dramatically cleans up the model. In particular, we now have a
*unique* `Context` for any given "world".