Simplify erasure of arrays by going through an intermediate structure.
We introduce an actual type `ErasedTypeRef` which corresponds to
an erased type, with a 1 to 1 correspondance to JVM types. It
still retains semantic information, in the form of a ClassSymbol
and possibly a number of array dimensions.

In a second step, we encode an `ErasedTypeRef` in the weird way
that `TermSig` wants.