TestPlatform.loadClasspath() now returns a Future[Classpath].
This will be necessary to implement its JavaScript counterpart.

It means that we must always asynchronously create a Context out
of that Classpath. MUnit `test` methods can accept bodies that
return `Future`s and correctly wait for them to be completed.

We leverage that in our existing `testWithContext`-like helpers:
they can deal with the asynchronous stuff, while the written tests
can keep being synchronous.

For SymbolSuite and PositionSuite, which did not yet have such
helpers, we add some.