Strictly separate TypeBounds from Types.
Before, `TypeAlias`es were *both* a `TypeBounds` and a `Type`.
In our system, this makes no sense.

We now make `TypeBounds` an `abstract class` to be certain that
nothing can every be both a `Type` and a `TypeBounds`.

Making `TypeAlias` not a `Type` anymore surfaces a number of places
where it was treated as a `Type`, most likely uncovering latent
issues.