The Great Symbols Refactoring.
Previously, the hierarchy of `Symbol` was limited to `ClassSymbol`,
`PackageSymbol` and `RegularSymbol`. `RegularSymbol`s could equally
represent terms or (non-class) types.

In the latter case, their `declaredType` would contain a dubious
encoding of the type definition. Notably, we used
`WildcardTypeBounds` as a way to wrap `TypeBounds` in a `Type`, but
that was a hack, and they had to be extracted everywhere.

Now, the hierarchy is more elaborate. In particular, we distinguish
`TermSymbol`s, with a `declaredType: Type`, from the `TypeSymbol`s,
which are either `ClassSymbol`s or `TypeSymbolWithBounds`, the
latter being further classified and exposing `bounds: TypeBounds`.

Some aspects of `Type`s are refined to expose and manipulate the
appropriate subtypes of `Symbol`s. In particular, `TermRef`s are
guaranteed to point to `TermSymbol`s while `TypeRef`s point to
`TypeSymbol`s.