Take a mutable Array in the constructor of TastyUnpickler.
This is more honest, as it then manipulates it everywhere as a
mutable array.

From `Loader`, we create a copy of the bytes as a mutable Array.
The original bytes come from a public-facing `IArray`, so we
really should not be mutating them, and this is the only way to
make sure of it.