Don't expose UTF-8 name creation in Names.
Instead, keep that in `reader`, which is the only place where this
matters.

We isolate the actual decoding in a new helper object `UTF8Utils`.
Since this object is the only place where we still cast an
`IArray[Byte]` to an `Array[Byte]`, we directly do it, and remove
`unsafe.asByteArray`.