Lock down any access to the `designator` of `NamedType`s.
The designator is not a stable value. Only `name` and `symbol`
should be accessed from the outside.

To hide the `designator`, we have to make `TermRef` and `TypeRef`
non-case classes.

For testing purposes, we expose a `private[tastyquery]` access to
the designator with a scary name.

These changes also allow to make `LookupIn` and
`Scala2ExternalSymRef` `private[tastyquery]`, which furthers hide
implementation details of the library.