Make `Type`s non-case classes.
`Type`s do not have a meaningful structural equality, and we
should not give the impression that they do. A number of `Type`s
were already non-case classes. We now make all of them non-case.

We introduce custom extractors for testing purposes in
`ReadTreeSuite`.

This change makes it much more obvious that something's up with
type reading in the Scala 2 unpickler. We have to tolerate reading
the same type address several times, so we remove one assertion.
We should revisit this one day.