Use a local-only TypeMap for `integrate` and `derived*` methods.
Before, all `TypeMap`s required a `Context` and performed some
amount of non-local processing, like normalizing derived
selections. Even the computation of derived `Binders` and
`integrate`, which should have been purely local computations,
could therefore access non-local information. Because of that,
we had to delay the initialization of `LambdaType`s.

We now make a distinction between a `TypeMap`, which uses no
`Context` and therefore performs a "syntactic"-only
transformation, from a `NormalizingTypeMap`, which uses a
`Context` to perform semantic normalizations.

The substitutions of `Binders` for `derivedLambdaType` and of
param `Symbol`s for `integrate` are adjusted to use a local-only
`TypeMap`. Similarly, all the `derived*` methods are amended to
be local-only, without `Context`. These changes must be done
together with the `TypeMap` split, since they are interdependent.

For `derived*` methods that actually performed semantic
normalization (notably `derivedSelect`), that behavior is moved to
a separate method, which does require a `Context`. Only the
`NormalizingTypeMap` calls those.