Make sure that all created Symbols receive Flags.
We introduce a generic method `Symbol.checkCompleted()`, which will
eventually check that all fields of a `Symbol` have been properly
initialized. Currently, it only checks that the `flags` are
initialized, which is common to all kinds of symbols.

We call that method at the end of reading a file, for all the
symbols that were created while reading that file. This makes sure
that by the time a `Symbol` gets in the hand of a user, it has been
completely initialized.

Adding those checks already identifies a few places where we did
not initialize the flags. We fix those in this commit as well.