Simplify how package declarations are created and looked up.
We move the logic to get-or-create a package decl to
`PackageSymbol`. This allows to simplify the logic and a stronger
guarantee that no one creates packages other than through the new
method `getPackageDeclOrCreate`.

We rename the remaining `Context.createPackageSymbolIfNew` method
to `findPackageFromRootOrCreate`, which makes it more consistent.

We give a stronger public guarantee to `getPackageDecl`, which,
combined with the above changes, allows us to use `getPackageDecl`
everywhere instead of `getPackageDeclInternal`.

This, in turns, allows to integrate `getPackageDeclInternal`
directly in `getPackageDecl`, further simplifying the design and
making it easier to reason about.