Simplify the interface of root completion in `Loaders`.
Instead of returning an `Option[Symbol]`, we only return a
`Boolean` indicating whether any root entry was successfully
loaded. It may be that we return `true` even though no *symbol*
was actually created.

From `Symbols`, we do a second attempt at `getDeclInternal` when
`loadRoot` returns `true`, instead of taking the result of
`loadRoot`.

With this change, `Loaders` does not need to use `getDeclInternal`
anymore.