Remove the `LazyLoadingSuite`.
It has outlived its usefulness. We have many more tests now that
would break if some symbols were loaded earlier than we want,
notably the ones with restricted classpaths.

One of the tests was `ignore`d because it was flaky anyway.

The other test's only remaining value was to test that, if we load
`pkg.SomeClass`, we don't trigger loading
`pkg.SomeOtherExistingClass`.

Removing `LazyLoadingSuite` allows us to contain `getDeclInternal`
to be `protected[this]` in `DeclaringSymbol`, which is a much
stronger guarantee that we do not abuse it anywhere.