Remove `ClassInfo`.
All the code paths involving `ClassInfo` were dead code. Now that
we have a significant portion of subtyping, I am confident that
its need will not resurface later.

This introduces an `AssertionError` path in `TypeRef.underlying`,
which is not nice, but it also removes some `???` code paths in
other places.

It would be worth investigating whether we could make `TypeRef`
not a `TypeProxy`, but a cursory attempt met other issues.