Remove `NoSymbol`; use an internal `NoExternalSymbolRef` in `PickleReader`.
`NoSymbol` was only used in one place at this point: when reading
a `NONEsym` reference in `PickleReader`. We now use an internal
`NoExternalSymbolRef` instead.

One apparently legitimate use of `NONEsym` in a Scala 2 pickle is
as the `owner` of an `ExternalSymbolRef` for the root package. In
that case, the result of `readMaybeExternalSymbolRef()` is ignored.

There is also a non-legitimate use, as the `owner` of some
`TYPEsym`s that are never used themselves. We work around that by
replacing those `TYPEsym`s with `NoExternalSymbolRef` as well.
Further investigation could happen later if that results in an
observable crash.