Add tree categories for term trees and pattern trees.
`TermTree`s are the only ones that have a generic `tpe`. Most term
trees accept only `TermTree`s as arguments. For example, the `fun`
and `args` or an `Apply` tree are all `TermTree`s.

Restricting `tpe` to `TermTree`s allows to get rid of many uses of
`NoType`.

The pattern of `CaseDef` is limited to `PatternTree`s. Some
"physical" TASTy tags are shared between similarly-looking term
trees and pattern trees. The unpickler tells them apart. This way,
when looking at patterns, the semantics are clearer.

The bindings of `Inlined` are limited to `ValOrDefDef`, a new
common superclass of `ValDef` and `DefDef`.