Remove `NoType`.
There was only remaining use of `NoType`, in the Scala 2 unpickler.
Unlike TASTy, the Scala 2 pickle format has a `NOtpe` tag. However,
it is unclear when that can be a valid type for any symbol.

We did find an occurrence of it in `scala.collection.MapView`, but
we could not investigate all the way to be able to ignore it. We
know enough to associate it with refinement types. Eventually, we
should not have symbols at all for refinement types, so it is
possible that we could completely remove that in the future.

In the meantime, we invent a synthetic type alias
`<notype> = Nothing` to be able to unpickle these Scala 2 corner
cases.