Handle polymorphic opaque type aliases.
In TASTy, their info is stored as a TypeLambda over "opaque
TypeBounds" (BoundedType's in our implementation). However, that
does not fit our model for `TypeMemberDefinition`.

We now distribute `TypeLambda`s over `BoundedType`s so that we can
correctly handle polymorphic opaque type aliases.

We do not handle variance yet.

This commit includes a fix to `dealias` not to follow opaque type
aliases. From a public point of view, it does not make sense to do
so.