Be even more tolerant when typing polymorphic `Apply(New, ...)`.
I was not able to reproduce the issue in a standalone test, but
apparently it is possible to have *both* a `New` with an
`AppliedType` (hence, monomorphic) wrapped in a `TypeApply`. This
would attempt to reapply type parameters to an already monomorphic
type, which crashed.

We are now a bit more tolerant, and apply type parameter lists to
the type of the `New` as long as we have a polymorphic type. We
stop applying further type parameter lists when the type is fully
monomorphic.