Make `signature`, `targetName` and `signedName` more user-friendly.
We make `signature` and `targetName` public. All terms now have
a `signature` (it is non-optional). For fields, it is the signature
of the JVM/IR getter method, which is also the descriptor of the
underlying JVM field.

In contrast, `signedName` does not return a `SignedName` for
parameterless methods anymore. It returns a `SignedName` if and
only if a `SignedName` is required to find the symbol through
`getDecl`. To access the JVM signature of a parameterless method,
`signature` can be used.

We move `NamedType.optSignature` to `TermRef` so that it can
become non-optional there as well.