Refactoring: Prepare the codebase for `TypeRef`s not to always have symbols.
The `NamedType.symbol` method is conceptually moved to
`TermRef.symbol`. Instead, we introduce `NamedType.optSymbol`,
and adapt the codebase to be ready to support `TypeRef`s that
do not have an underlying `symbol`.

In this commit, we actually keep `NamedType.symbol` in the API
to separate API changes from the internal refactorings.

We add some more convenience helpers for detecting and dealing
with `TypeRef`s that point to classes. Those will always have a
symbol which is a `ClassSymbol`.

Opaque type aliases will always have a symbol, but other kinds of
type members will not always have one. So when we need to access
bounds, we use dedicated new accessors in `TypeRef`.