Introduce public member lookup methods returning Symbols in ClassSymbol.
Instead of having to through `cls.typeRef`.

We make `Type.member` private. Since later, selecting a member on
a `Type` will not always return a `Symbol` anymore, this method
will not make sense. Users can still construct a `NamedType` out
of a `Type` and `Name` (notably with the `select` methods) and
query information through it.