Delay the re-resolution of members in normalizedDerivedSelect.
When the prefix of a `NamedType` changes, in some cases we must
re-resolve the symbol, as it can be specialized in a subclass (or
later in a refinement type).

Previously, we did this immediately during
`normalizedDerivedSelect` to get a new symbol. Now, we only
preserve the (signed) name, and delay the re-resolution to when
the produced `NamedType` will need it.

We must preserve the symbol if it is `Private`, since re-resolving
it by name would be incorrect.

This change also removes more uses of `Type.member` to get to a
symbol, which will not be possible with the coming changes.