Member resolution from Type that can handle multiple- or no symbols.
Instead of having `Type.findMember` which returns an
`Option[Symbol]`, we now have a `resolveMember` method that
returns a `ResolveMemberResult`. A `ResolveMemberResult` retains
all the possible target symbols (that are not overridden by
another one), along with the resolved, `asSeenFrom`ed type or
type bounds. For type members, the list of symbols can be empty.

`TypeRef` and `TermRef` now use this method, and rely on the
parts of `ResolveMemberResult` to provide underlying types, type
bounds, etc., instead of relying on a unique resolved `Symbol`.

In this commit, the list of symbols is always non-empty, but it
will be possibly empty when we add support for type refinements.