Full representation of match types.
`MatchTypeCase` now directly represents captures in the form of
type parameters. Previously, a capture match-case would be
represented as a `TypeLambda` over a `MatchTypeCase`. That was
inherited from how they are stored in TASTy, but it does not make
much sense, since we cannot actually use the `TypeLambda` in the
way as real ones.

`MatchTypeCase` is not a `Type` anymore. It can only happen in
the `cases` of a `MatchType`.

We also implement `MatchTypeTree.calculateType`.