Short-circuit no-op `asSeenFrom(tp, C.this, C)` to `tp`.
By definition, `asSeenFrom(tp, C.this, C)` means substituting
`C.this` for `C.this` inside `tp`, so it is a no-op.

We now explicitly short-circuit it. Previously, it would cause
infinite recursion in `baseTypeOf` when parent types have
higher-kinded type parameters that are instantiated in a subclass
to the subclass itself.