Cache the core, recursive aspect of `baseTypeOf`.
Eventually, every `cls.baseTypeOf(tp)` recurses to a point where
`tp` is `clsOwner.this.cls`. That computation is expensive,
because it recurses into all the base classes of `cls` until it
finds `this`.

Therefore, we now cache that particular step, which is safe to do
because it only depends on the pair of `ClassSymbol`s `(this, cls)`.