Simplify baseTypeOf.
We make the "terminal" cases of `baseTypeOf` more explicit: the
cases where `tp` is `prefix.C[...Ts]` (where `...Ts` can be empty).

Previously, there were more recursions through `recur` at the level
of `AppliedType` and `TypeRef.OfClass` when it was a class but not
`this` class.

Now, there are two main cases:

* If it is `prefix.C[...Ts]`, finish with `baseTypeOfClassTypeRef`
  and substitute the `...Ts` (if any).
* Otherwise, recurse through `baseTypeOf(tp.superType)`.

We also make the remaining cases more exhaustive, instead of
relying on a silent `case _ =>`.