Translate `Seq[T] @Repeated` to `<repeated>[T]` in Scala 3 `MethodType`s.
This aligns how repeated parameters are represented in `MethodType`s
for Scala 3, Scala 2 and Java. It also aligns those types with the
call-site type of the `Typed` nodes.

The `ValDef` for the parameter is still typed as `Seq[T] @Repeated`,
since inside the method it must be treated as a `Seq`.