Handle subtyping and member selection of TermRefinement with MethodicType.
When a `TermRefinement` has a `MethodType` or a `PolyType`, we
cannot use `Type.resolveMember(refinedName)`. `resolveMember` needs
the correctly *signed* name of such terms. We cannot compute the
fully correct signed name because the result type of the target
member might be a subtype (with a different erasure) of the
requested `refinedType`.

In order to handle this situation, we add a new method
`Type.resolveMatchingMember`, which matches the simple name and
the parameters of the signature, but not its result. We compensate
by adding a customizable predicate on the type. The implementation
of `resolveMatchingMember` is unfortunately a bit duplicated from
`resolveMember`.

We can get into `resolveMatchingMember` from two places: in
subtyping when the right-hand-side is a `TermRefinement`, or from
`TermRefinement.resolveMember` itself.

Because of subtyping, we must be able to resolve term members that
do not have any accompanying `TermSymbol`. So
`ResolvedMemberResult.TermMember` can now be constructed with an
empty list of `TermSymbol`s. However, if such a result reaches a
`TermRef` (possibily via `lookupMember`), we consider it as
`NotFound` instead.