Take the owner ref of `TYPEREFin` into account.
`TYPEREFin` is used to persist references to private type members
in some very specific circumstances. The typical case is when
persisting the result of capture conversion, since we need to
select a class type parameter of an external class. It can also
happen to select shadowed class members (from Java).