Clean up the way we read recursive types.
For `RecType`/`RecThis`, use the infrastructure for enclosing
binders in the `localCtx`, like we do for `LambdaType`s.

For the `THIS` references to the enclosing refinement class, we
now create the magic refinement class symbols along with all the
other symbols, and then we specially treat `REFINEDtpt` in
`readType` position to fetch that symbol.