Handle more cases of capture conversion (hopefully all).
If the widened type of a SingletonType is not directly an applied
class type (e.g., it is an applied *alias* to a class type),
capture conversion would not trigger. There would be two steps
required to go from the SingletonType to the correct AppliedType
that can be matched in `isMatchingApply`, and the first step would
hide the SingletonType from a potential capture conversion.

Instead, we now integrate capture conversion at the time of
computing the base type of the left-hand-side. This is the only
place where we actually retry the process with a widened
left-hand-side, except for the last-resort call to
`singletonType.underlying` in `level4`. Most likely, if we use the
last resort, capture conversion would never help.