Remove the `Context` argument from `TypeTree.toType` and similar.
`toType` is used on arbitrary `TypeTree`s during unpickling.
Therefore, it must be able to complete without resolving arbitrary
types. The same goes for `TypeTree.toPrefix`,
`TypeArgTree.toTypeOrWildcard` and `TypeBoundsTree.toTypeBounds`.

This change surfaces a difficulty for `SingletonTypeTree` and
`TermRefTypeTree`. Indeed, they contain a `TermTree` from which
they need the `tpe`. `TermTree.tpe` does require a `Context`, as
it must resolve external references in the general case.

To address this issue, we introduce an alternative method
`TermTree.toTypePrefix`. It does not require a `Context`, but can
only succeed for term trees that constitute valid type prefixes,
i.e., paths.