Deprecate `TermType.widen`.
Widening in general is not an operation supported by the spec. It
is somewhat useful in internal algorithms, but can usually be
replaced by more targeted type tests.

The main use case for widening is to get to the underlying type of
a term ref. For example, when applying a method, we get a `TermRef`
that *references* the target method, but in terms of *type*, we
want its underlying. Likewise, when assigning to a variable, the
rhs has to comply to the underlying type of the target variable,
not its singleton type.

We provide that operation as `widenTermRef`. It is more precise
than `widen` for these use cases, as it won't dereference a term
ref *twice*, in case its underlying is itself a `TermRef`, i.e.,
a singleton type of another `val`.