Introduce a more restrictive `ReaderContext` inside `reader.*`.
It statically makes sure that we never access anything that
requires reading another file.

This change surfaced a couple places where there could potentially
be external reading, and which are now a bit more complicated to
fit the static restriction. The most prominent example is
`TreeUnpickler.readWithin`, which also suggested that we decouple
reading the `privateWithin` from the flags (to only do it after
`createSymbols()` is over).