Use FunctionN.apply as the resolved `symbol` for poly function `apply`s.
Normally, a *term* selection *always* points to a symbol, at the
TASTy level. If we select a term member that only exists through
refinements at the source level, it gets expanded into a
`Selectable`-based desugaring.

There is however *one* exception (also visible in the spec): we
can select a method member named `apply` without a symbol if the
type extends `scala.PolyFunction`.

This puts us in a difficult situation. Regressing `TermRef.symbol`
and `TermReferenceTree.symbol` into an optional `optSymbol` just
for that rare special-case can be very annoying. At the same time,
we cannot afford to throw.

To solve this dilemma, we choose to resolve to the corresponding
`FunctionN.apply` symbol. This is more meaningful than it looks:
it is the method that will actually get called at run-time, and
it makes for a good "go to definition" target.