Only use TermName's in signatureName's.
Previously, there was an invariant that the last name in a
signature name was always a `TypeName`, and all the other ones
were `TermName`s. Instead, we now declare that they must all be
`TermName`s.

This removes the need for a number of `mapLast`s, and simplifies
manipulations and invariants.