Move `Symbol.fullName` to `PackageSymbol`; introduce `Symbol.displayFullName`.
`fullName` is not meaningful on arbitrary symbols. Having it
return a supposedly structured `FullyQualifiedName` is misleading.

We move it `PackageSymbol`, for which it is actually meaningful and
used.

In `Symbol`, we introduce `displayFullName` instead, which does
the same as what the previous `fullName` did but returned as a
`String`. This is meant to be used for display purposes only, such
as debugging or error messages.