Move `Contexts.init` to `Contexts.Context.initialize`.
The way our library's API is organized, it is common to import
the contents of top-level objects with wildcard imports. For
`Contexts` however, doing `import tastyquery.Contexts.*` would
pollute the scope with a bare identifier `init`, which is not
great.

Instead, users will now have to write
`Context.initialize(classpath)`.