Source-in what we need from tasty-core.
We already had a source dependency on tasty-core because of the
Scala.js artifact. This means that things from the tasty-core
artifact leaked in our public API.

Depending on the upstream tasty-core, while somewhat convenient,
also constrained us to only be able to read TASTy from a published
version of dotc.

We now source-in what we need from tasty-core (which is not that
much) as package-private.