Implement reading annotations from Scala 2 pickles.
We do not read all kinds of trees. Only the ones that do not
contain definitions nor pattern matches (which means no `try`s
either). If we encounter a kind of tree we do not support, we bail
and replace the annotation argument with a placeholder.

We have to cheat for the selection of the constructor because the
pickles do not contain a signed name reference. However, we have
enough information for the most important methods of `Annotation`
to actually work.