Merge the functionality of Descriptors into JavaSignatures.
With the exception of fields of a base type, descriptors are a
strict subset of member signatures. We generalize JavaSignatures
to be able to read such field descriptors, so that we can remove
Descriptors in favor of JavaSignatures.

We cannot do the same for *class* signatures, unfortunately,
because monomorphic extends clauses are store in a completely
different way that polymorphic ones.