Decouple detecting class kinds from storing specific attributes.
First, we read the structure of a classfile, independently of its
class kind.

Then, for a top-level class, we detect its class kind. At that
point, we only look at the `Scala`, `ScalaSig` and `TASTY`
attributes. We do not store any reference to other specific
attributes.

Third, the Loaders switch on the kind to use the appropriate
loading functions. The loading function for Scala 2 rescans the
attributes to find the signature annotation. The loading function
for Java creates a map of all the available attributes, from which
it uses--for now--the `InnerClasses` and `Signature` attributes.

This frees the `Loaders` from manipulating specific sets of
attributes that are needed by the loading functions. It will also
be necessary for reading Java annotations, since they are stored
in more attributes than were currently identified.