Fix #397: Use a fake owner for orphan symbols in Scala 2 pickles.
We had a long-standing issue with "orphan" `TYPEsym`s and `VALsym`s
in Scala 2 pickles. We were replacing them by "no symbol"
themselves, because we did not know what to do with them.

Issue #397 finally provided a good setup to diagnose it. It turns
out that those orphan symbols can be found as term and type params
of method types and poly types, in particular, Scala 2's encoding
of type lambdas, which can be produced by kind-projector.

Since we now have a good reason to have those orphan symbols, we
bite the bullet and allocate one global symbol to "host" all those
symbols coming from Scala 2.