Fix #398: Include inherited members in the PackageRef fallback lookup.
If we have a `package object` that extends other classes, we may
have to look member selections among inherited members. This is
done by `getMember` instead of `getDecl`.

To test this, we need a Scala 2 `package object` with an `extends`
clause. There is no such thing in the standard library, which is
our only source of Scala 2 code, unfortunately. So we don't add a
test for that.