Fix #401: Avoid asSeenFrom for types/bounds that don't depend on the prefix.
This mostly acts as a fast path. However, for the particular case
of static recursive match types, it more critically cuts some
infinite recursion.

Normally, when a match type refers to itself, it would do so
through a `ThisType`. Computing `asSeenFrom` for a `ThisType`
prefix already cuts the infinite recursion because it is, by
construction, a no-op.

However, for static recursive match types, the compiler makes them
refer to themselves through their public, static prefix. This
caused the infinite recursion observed in #401.

To avoid the issue, we use another shortcut to `asSeenFrom`: if
the type we map over does not contain any `ThisType`, we also
know it will be a no-op. We store that as a lazy val
`isPrefixDependent` of `TermSymbol` and `TypeMemberSymbol`, so
that they can take a short cut in `{type,bounds}AsSeenFrom`.