Pickle reader: Read the typeParams of classes in `completeSymbolType`.
We only need the `typeParams` when completing the type of
constructors, and the types of constructors are not needed by the
class. Therefore, we can move reading `typeParams` until completing
the type of the class.

This simplifies things, because previously we were doing a sort of
look-ahead inside the `POLYtpe`, which also required having a
special non-caching version of `at()`. Now we can genuinely read
the `POLYtpe` and extract the type parameters from it after the
fact.