Pickle reader: Handle name and flags special-casing per symbol kind.
Previously, we tried to write "generic" code for handling all the
special cases to apply to names and flags. However, that was making
the logic harder than necessary, with too many possible branching
paths.

We now move that logic inside the various `case`s of `tag match`.
This makes the logic simpler overall.